(redesign/new-transaction) - fix file skip logic during new transaction creation
